Move duplicated code to super class
