+++ restapi/src/main/java/bisq/restapi/RestApi.java
@@ -19,41 +19,22 @@ package bisq.restapi;
 
 
 import bisq.core.account.witness.AccountAgeWitnessService;
-import bisq.core.app.TorSetup;
 import bisq.core.app.misc.ExecutableForAppWithP2p;
 import bisq.core.dao.SignVerifyService;
 import bisq.core.dao.governance.bond.reputation.BondedReputationRepository;
 import bisq.core.dao.governance.bond.role.BondedRolesRepository;
 import bisq.core.dao.state.DaoStateService;
 import bisq.core.dao.state.DaoStateSnapshotService;
-import bisq.core.user.Cookie;
-import bisq.core.user.CookieKey;
 import bisq.core.user.Preferences;
-import bisq.core.user.User;
 
-import bisq.network.p2p.P2PService;
-import bisq.network.p2p.P2PServiceListener;
-import bisq.network.p2p.peers.PeerManager;
-
-import bisq.common.Timer;
-import bisq.common.UserThread;
 import bisq.common.app.Version;
-import bisq.common.config.BaseCurrencyNetwork;
 import bisq.common.config.Config;
-import bisq.common.handlers.ResultHandler;
-
-import com.google.inject.Key;
-import com.google.inject.name.Names;
 
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
-//todo not sure if the restart handling from seed nodes is required
 
 @Slf4j
 public class RestApi extends ExecutableForAppWithP2p {
-    private static final long CHECK_CONNECTION_LOSS_SEC = 30;
-
-    private Timer checkConnectionLossTime;
     @Getter
     private DaoStateService daoStateService;
     @Getter
@@ -64,6 +45,8 @@ public class RestApi extends ExecutableForAppWithP2p {
     private BondedRolesRepository bondedRolesRepository;
     @Getter
     private SignVerifyService signVerifyService;
+    private DaoStateSnapshotService daoStateSnapshotService;
+    private Preferences preferences;
 
     public RestApi() {
         super("Bisq Rest Api", "bisq_restapi", "bisq_restapi", Version.VERSION);
@@ -80,113 +63,31 @@ public class RestApi extends ExecutableForAppWithP2p {
         checkMemory(config, this);
     }
 
-
-    ///////////////////////////////////////////////////////////////////////////////////////////
-    // We continue with a series of synchronous execution tasks
-    ///////////////////////////////////////////////////////////////////////////////////////////
-
     @Override
     protected void applyInjector() {
         super.applyInjector();
 
-        injector.getInstance(DaoStateSnapshotService.class).setResyncDaoStateFromResourcesHandler(this::gracefulShutDown);
+        preferences = injector.getInstance(Preferences.class);
+        daoStateService = injector.getInstance(DaoStateService.class);
+        accountAgeWitnessService = injector.getInstance(AccountAgeWitnessService.class);
+        bondedReputationRepository = injector.getInstance(BondedReputationRepository.class);
+        bondedRolesRepository = injector.getInstance(BondedRolesRepository.class);
+        signVerifyService = injector.getInstance(SignVerifyService.class);
+        daoStateSnapshotService = injector.getInstance(DaoStateSnapshotService.class);
     }
 
     @Override
     protected void startApplication() {
         super.startApplication();
 
-        Cookie cookie = injector.getInstance(User.class).getCookie();
-        cookie.getAsOptionalBoolean(CookieKey.CLEAN_TOR_DIR_AT_RESTART).ifPresent(cleanTorDirAtRestart -> {
-            if (cleanTorDirAtRestart) {
-                injector.getInstance(TorSetup.class).cleanupTorFiles(() ->
-                                cookie.remove(CookieKey.CLEAN_TOR_DIR_AT_RESTART),
-                        log::error);
-            }
-        });
-
-        injector.getInstance(Preferences.class).setUseFullModeDaoMonitor(false);
-
-        daoStateService = injector.getInstance(DaoStateService.class);
-        accountAgeWitnessService = injector.getInstance(AccountAgeWitnessService.class);
-        bondedReputationRepository = injector.getInstance(BondedReputationRepository.class);
-        bondedRolesRepository = injector.getInstance(BondedRolesRepository.class);
-        signVerifyService = injector.getInstance(SignVerifyService.class);
-
-        injector.getInstance(P2PService.class).addP2PServiceListener(new P2PServiceListener() {
-            @Override
-            public void onDataReceived() {
-                // Do nothing
-            }
-
-            @Override
-            public void onNoSeedNodeAvailable() {
-                // Do nothing
-            }
-
-            @Override
-            public void onNoPeersAvailable() {
-                // Do nothing
-            }
-
-            @Override
-            public void onUpdatedDataReceived() {
-                // Do nothing
-            }
-
-            @Override
-            public void onTorNodeReady() {
-                // Do nothing
-            }
-
-            @Override
-            public void onHiddenServicePublished() {
-                boolean preventPeriodicShutdownAtSeedNode = injector.getInstance(Key.get(boolean.class,
-                        Names.named(Config.PREVENT_PERIODIC_SHUTDOWN_AT_SEED_NODE)));
-                if (!preventPeriodicShutdownAtSeedNode) {
-                    startShutDownInterval();
-                }
-                UserThread.runAfter(() -> setupConnectionLossCheck(), 60);
-
-                accountAgeWitnessService.onAllServicesInitialized();
-            }
-
-            @Override
-            public void onSetupFailed(Throwable throwable) {
-                // Do nothing
-            }
-
-            @Override
-            public void onRequestCustomBridges() {
-                // Do nothing
-            }
-        });
-    }
-
-    private void setupConnectionLossCheck() {
-        // For dev testing (usually on BTC_REGTEST) we don't want to get the seed shut
-        // down as it is normal that the seed is the only actively running node.
-        if (Config.baseCurrencyNetwork() == BaseCurrencyNetwork.BTC_REGTEST) {
-            return;
-        }
-
-        if (checkConnectionLossTime != null) {
-            return;
-        }
-
-        checkConnectionLossTime = UserThread.runPeriodically(() -> {
-            if (injector.getInstance(PeerManager.class).getNumAllConnectionsLostEvents() > 1) {
-                // We set a flag to clear tor cache files at re-start. We cannot clear it now as Tor is used and
-                // that can cause problems.
-                injector.getInstance(User.class).getCookie().putAsBoolean(CookieKey.CLEAN_TOR_DIR_AT_RESTART, true);
-                shutDown(this);
-            }
-        }, CHECK_CONNECTION_LOSS_SEC);
-
+        preferences.setUseFullModeDaoMonitor(false);
+        daoStateSnapshotService.setResyncDaoStateFromResourcesHandler(this::gracefulShutDown);
     }
 
     @Override
-    public void gracefulShutDown(ResultHandler resultHandler) {
-        super.gracefulShutDown(resultHandler);
+    protected void onHiddenServicePublished() {
+        super.onHiddenServicePublished();
+
+        accountAgeWitnessService.onAllServicesInitialized();
     }
 }
